stopwords.txt file management to play nice with Docker automated builds
Docker automated builds don't allow to trigger actions before the build.
So, in the previous commit, as we needed to do "make" between to run a build,
it weren't possible for Docker registry to autobuild that.

The idea behind stopwords.txt were to allow a Makefile to decide how to
build this file, from various source including Phabricator (but in the future
extended to other common words specific to an application to ignore).

We are currently including this file in the repository, so we can get the
benefit of and customize the file directly in this repository.